Implement a "compat" entry point that falls back to the uSES shim
We use the same trick we do with batchedUpdates: a module with
variables + getter/setter functions. Each entry point is then
responsible for importing the right variables and setting them.
This happens before the modules are loaded, so we can then import
the getters where they're needed and call them to get the right
instances of `useSyncExternalStore`.

The "compat" and "alternate-renderer" entry points both use the
`uSES` shim, so they should be compatible back to React 16.9+.
The primary entry point imports directly from React, and thus
assumes React 18.